Added a number of servings to the recipe object, and removing the subject from the comment object.
Also, removed the overridden setGender method as it was causing
minor issues with data loading.  Will think about how to implement a bool to text
interface for this later.  Added a @todo in lib/model/User.php for this purpose.

Will continue to add data to the fixtures.  Have been finding (of all things) 
http://www.marthastewart.com/recipe/best-tortilla-and-black-bean-pie?backto=true
to be a useful resource in how to lay this out.  Although, this project will
rapidly diverge from the simple interface she provides.