Found the issue with our comments not loading.  Apparently, propel has a class called Comment, and we can't have a conflict with this name.  I've changed all references of Comment to Coment (note the lack of one of the m's.  Should work from now on.
I also added stuff to the fixtures and we can now control how many
test recipes and comments we have.  I've defaulted it to 250 recipes
and 20 comments per recipe.  Other additions to the fixtures have
and should continue to be made.

Also, implemented a simple rating system.  Will have to see if it
is too simple, but I think it will work.  For the rating system
look in lib/model/Recipe.php.

The recipe I added to the fixtures came from this very helpful site:
http://allrecipes.com/Recipe/Monicas-He-Man-Chili/Detail.aspx